Fix Profile implementation
Optional + Computed maps are not really possible in TF.
The current implementation supports the ability to overwrite/set
specific permissions but does not read them all back (computed).
This could be added later with a computed only map for readonly
purposes. The current import API also doesn't provider the user config
and therefore we can't import the permissions set in config during import